plugin compilerMetadata
