+++ libs/remix-core-plugin/compiler-content-imports/src/lib/compiler-content-imports.ts
@@ -24,7 +24,7 @@ export class CompilerImports extends Plugin {
 
   async setToken () {
     const protocol = typeof window !== 'undefined' && window.location.protocol
-    const token = await this.call('settings', 'getGithubAccessToken')
+    const token = await this.call('settings', 'get', 'settings/gist-access-token')
     this.urlResolver.setGistToken(token, protocol)
   }
