+++ libs/remix-core-plugin/compiler-metadata/README.md
@@ -1,0 +1,3 @@
+# remix-core-plugin-compiler-metadata
+
+This library was generated with [Nx](https://nx.dev).
