+++ libs/remix-core-plugin/compiler-metadata/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "@remix-core-plugin/compiler-metadata",
+  "version": "0.0.1",
+  "description": "This library was generated with [Nx](https://nx.dev).",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "Remix Team",
+  "license": "ISC"
+}
