+++ libs/remix-core-plugin/compiler-metadata/src/index.ts
@@ -1,0 +1,1 @@
+export * from './lib/compiler-metadata'
